+++ Magic/CHANGELOG.md
@@ -8,6 +8,7 @@
  - Add /mage clear command for clearing player inventories of magic items
  - Add default_mage_class config option
  - Improve and optimize Citizens integration, target Sentinel NPCs by default
+ - Add ModifyMaxHealth action
 
 # 7.3.7